Fix Tiny Unicode and Short Unicode byte length ranges
I'm running `pysmile` 0.2 and I'm seeing Tiny Unicode strings being
encoded slightly different to what the specification describes.

Encoding a UTF-8 string consisting of 4 characters (i.e. the string
`data`) is encoded as follows:

```
83 6461 7461
```

The hexadecimal string `6461 7461` equals `data`. The prefix is 0x83.

The specification describes that 0x80 would encode a UTF-8 string with 1
character, however this is not allowed as a single-byte character string
must be an ASCII string, and therefore the ASCII encoding is preferred.
Therefore 0x81 describes a UTF-8 string with 2 bytes, 0x82
describes a UTF-8 string with 3 bytes, and of course 0x83 maps to a
UTF-8 string with 4 bytes, which is what I get when encoding my string
with `pysmile`.

The problem is that following that logic, the LSB of the Tiny Unicode
prefix equals the byte-length of the string - 1 (i.e. 0x83 = `10000011`,
the 5-bit LSB is 3, so the byte length is 4). The upper bounds of the
Tiny Unicode range is 0x9f = `10011111`, where the 5-bit LSB is 31. This
maps to a 32 byte UTF-8 string, so the largest UTF-8 string that the
Tiny Unicode encoding can handle is NOT 33 bytes as described by the
specification, but 32.

As a consequence, the Short Unicode encoding starts from 33 and not from
34.

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>